+++ src/container/freelancer/SavedJobs.js
@@ -12,6 +12,7 @@ import { GetViewAllSavedJobsAction } from '../../redux/Freelancer/FreelancerActi
 import axios from 'axios'
 import { useEffect } from 'react'
 import { toast } from 'react-toastify'
+import { timeAgo } from './TimeFunctions'
 
 const SavedJobs = () => {
 
@@ -66,34 +67,6 @@ const SavedJobs = () => {
         }
     };
     
-
-   
-    function timeAgo(postedTimeStr) {
-        const postedTime = new Date(postedTimeStr);
-        const currentTime = new Date();
-
-        const deltaInMilliseconds = currentTime - postedTime;
-        const deltaInSeconds = Math.floor(deltaInMilliseconds / 1000);
-        const deltaInMinutes = Math.floor(deltaInSeconds / 60);
-        const deltaInHours = Math.floor(deltaInMinutes / 60);
-        const deltaInDays = Math.floor(deltaInHours / 24);
-
-        if (deltaInMinutes < 1) {
-            return "just now";
-        } else if (deltaInMinutes < 60) {
-            return `${deltaInMinutes}m ago`;
-        } else if (deltaInHours < 24) {
-            return `${deltaInHours}h ago`;
-        } else if (deltaInDays < 30) {
-            return `${deltaInDays}d ago`;
-        } else if (deltaInDays < 365) {
-            const months = Math.floor(deltaInDays / 30);
-            return `${months} month ago`;
-        } else {
-            const years = Math.floor(deltaInDays / 365);
-            return `${years} year ago`;
-        }
-    }
 
 
   return (
