+++ src/container/freelancer/ViewMoreProjectDetail.js
@@ -5,7 +5,7 @@ import frame from '../../components/images/Frame.png'
 import money from '../../components/images/money.png'
 import rating from '../../components/images/superstart.png'
 import { Link, useLocation, useNavigate } from 'react-router-dom'
-
+import { timeAgo } from './TimeFunctions';
 import { useDispatch, useSelector } from 'react-redux'
 import axios from'axios'
 import LightbulbOutlinedIcon from '@mui/icons-material/LightbulbOutlined';
@@ -63,33 +63,6 @@ const navigate = useNavigate();
     //     dispatch(GetFreelancerSelfBidAction(accessToken))
     // }, [])
   
-  function timeAgo(postedTimeStr) {
-    const postedTime = new Date(postedTimeStr);
-    const currentTime = new Date();
-  
-    const deltaInMilliseconds = currentTime - postedTime;
-    const deltaInSeconds = Math.floor(deltaInMilliseconds / 1000);
-    const deltaInMinutes = Math.floor(deltaInSeconds / 60);
-    const deltaInHours = Math.floor(deltaInMinutes / 60);
-    const deltaInDays = Math.floor(deltaInHours / 24);
-  
-    if (deltaInMinutes < 1) {
-        return "just now";
-    } else if (deltaInMinutes < 60) {
-        return `${deltaInMinutes} minute ago`;
-    } else if (deltaInHours < 24) {
-        return `${deltaInHours} hour ago`;
-    } else if (deltaInDays < 30) {
-        return `${deltaInDays} day ago`;
-    } else if (deltaInDays < 365) {
-        const months = Math.floor(deltaInDays / 30);
-        return `${months} month ago`;
-    } else {
-        const years = Math.floor(deltaInDays / 365);
-        return `${years} year ago`;
-    }
-  }
-
   const toggleSaveProject = async (project) => {
     try {
         let response;
@@ -146,7 +119,7 @@ const handleClick = (event,project) => {
                     <h1 className='text-xl font-normal font-cardo'>{project.title}</h1>
                     <p className='mt-4 text-base font-normal font-cardo'>{project.category}</p>
                     <div className='flex flex-row mt-2'>
-                        <div className=' basis-6/12'><p className=' font-inter font-normal text-base text-[#797979]'>Posted in {timeAgo(project.project_creation_date)}</p></div>
+                        <div className=' basis-6/12'><p className=' font-inter font-normal text-base text-[#797979]'>Posted {timeAgo(project.project_creation_date)}</p></div>
                         <div className=' basis-6/12'>
                             <div className=' text-right font-inter text-base font-normal opacity-[50%]'><i class="bi bi-geo-alt"></i>  Worldwide</div>
                         </div>
